Try leaving off a quote character. What happens? Was the error message helpful? Did it give any clues about what went wrong?
SyntaxError: invalid syntax. Perhaps you forgot a comma. It happens when we haven't followed variable name syntax/ rules. Also even if we follow syntax and resolve it by naming it as (Hello_Sid), we need to add quotes while printing. Which indicates it's a string to print, if not it takes it has variable and we need to declare or refer variable before using print method.